fix tests files
